requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
cloudinary==1.32.0
dj-database-url==0.5.0
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
oauthlib==3.2.2
phonenumberslite==8.13.7
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
